backend/requirements.txt
aiofiles==25.1.0
aiosqlite==0.19.0
asyncpg==0.29.0
cryptography==41.0.7
fastapi==0.104.1
mysql-connector-python==8.2.0
openai>=2.8.0
psycopg2-binary==2.9.9
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.6
uvicorn==0.24.0
uvicorn[standard]==0.24.0

# Data warehouse connectors
databricks-sql-connector==3.0.1
oracledb==2.0.0
pyodbc==5.0.1
google-cloud-bigquery==3.14.0
teradatasql==17.20.0.28
snowflake-connector-python==3.10.0
boto3==1.34.74

# Utilities
reportlab==4.0.7
xlsxwriter==3.1.9
db-dtypes
numpy<2
httpx>=0.25.0
